refactoring Message.ts with create init_vertexai.ts
